PIVY-89 Movies search by title, year, rating, my rating

+++ AnticevicApi/src/AnticevicApi.Model/Database/Main/User/Movie.cs
@@ -12,7 +12,7 @@ namespace AnticevicApi.Model.Database.Main.User
 
         public DateTime Timestamp { get; set; }
 
-        public int MyRating { get; set; }
+        public short MyRating { get; set; }
 
         public decimal Rating { get; set; }
